Conflicts:
	build/build-cloud.xml

+++ server/src/com/cloud/network/lb/LoadBalancingRulesManagerImpl.java
@@ -30,6 +30,7 @@ import javax.naming.ConfigurationException;
 
 import org.apache.log4j.Logger;
 
+import com.cloud.api.commands.CreateLoadBalancerRuleCmd;
 import com.cloud.api.commands.ListLoadBalancerRuleInstancesCmd;
 import com.cloud.api.commands.ListLoadBalancerRulesCmd;
 import com.cloud.api.commands.UpdateLoadBalancerRuleCmd;
@@ -42,6 +43,7 @@ import com.cloud.event.EventTypes;
 import com.cloud.event.UsageEventVO;
 import com.cloud.event.dao.EventDao;
 import com.cloud.event.dao.UsageEventDao;
+import com.cloud.exception.InsufficientAddressCapacityException;
 import com.cloud.exception.InvalidParameterValueException;
 import com.cloud.exception.NetworkRuleConflictException;
 import com.cloud.exception.PermissionDeniedException;
@@ -49,20 +51,21 @@ import com.cloud.exception.ResourceUnavailableException;
 import com.cloud.network.IPAddressVO;
 import com.cloud.network.LoadBalancerVMMapVO;
 import com.cloud.network.LoadBalancerVO;
-import com.cloud.network.Network;
 import com.cloud.network.Network.Service;
 import com.cloud.network.NetworkManager;
+import com.cloud.network.NetworkVO;
 import com.cloud.network.dao.FirewallRulesCidrsDao;
 import com.cloud.network.dao.FirewallRulesDao;
 import com.cloud.network.dao.IPAddressDao;
 import com.cloud.network.dao.LoadBalancerDao;
 import com.cloud.network.dao.LoadBalancerVMMapDao;
+import com.cloud.network.dao.NetworkDao;
 import com.cloud.network.lb.LoadBalancingRule.LbDestination;
+import com.cloud.network.rules.FirewallManager;
 import com.cloud.network.rules.FirewallRule;
 import com.cloud.network.rules.FirewallRule.Purpose;
 import com.cloud.network.rules.FirewallRuleVO;
 import com.cloud.network.rules.LoadBalancer;
-import com.cloud.network.rules.PortForwardingRuleVO;
 import com.cloud.network.rules.RulesManager;
 import com.cloud.user.Account;
 import com.cloud.user.AccountManager;
@@ -122,6 +125,14 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
     UsageEventDao _usageEventDao;
     @Inject
     FirewallRulesCidrsDao _firewallCidrsDao;
+    @Inject
+    FirewallManager _firewallMgr;
+    @Inject
+    ElasticLoadBalancerManager _elbMgr;
+    @Inject
+    NetworkDao _networkDao;
+    @Inject
+    FirewallRulesDao _firewallDao;
 
     @Override
     @DB
@@ -214,7 +225,7 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
             throw new InvalidParameterException("Invalid load balancer value: " + loadBalancerId);
         }
 
-        _accountMgr.checkAccess(caller.getCaller(), loadBalancer);
+        _accountMgr.checkAccess(caller.getCaller(), null, loadBalancer);
 
         try {
             loadBalancer.setState(FirewallRule.State.Add);
@@ -287,7 +298,7 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
             throw new InvalidParameterValueException("Unable to find load balancer rule " + loadBalancerId);
         }
 
-        _accountMgr.checkAccess(caller, rule);
+        _accountMgr.checkAccess(caller, null, rule);
 
         return deleteLoadBalancerRule(loadBalancerId, apply, caller, ctx.getCallerUserId());
     }
@@ -326,7 +337,7 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
         }
 
         txn.commit();
-
+        
         if (apply) {
             try {
                 if (!applyLoadBalancerConfig(loadBalancerId)) {
@@ -340,40 +351,19 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
         }
 
         _rulesDao.remove(lb.getId());
+        _elbMgr.handleDeleteLoadBalancerRule(lb, callerUserId, caller);
         s_logger.debug("Load balancer with id " + lb.getId() + " is removed successfully");
         return true;
     }
 
     @Override @DB
     @ActionEvent(eventType = EventTypes.EVENT_LOAD_BALANCER_CREATE, eventDescription = "creating load balancer")
-    public LoadBalancer createLoadBalancerRule(LoadBalancer lb) throws NetworkRuleConflictException {
+    public LoadBalancer createLoadBalancerRule(CreateLoadBalancerRuleCmd lb,  boolean openFirewall) throws NetworkRuleConflictException, InsufficientAddressCapacityException {
         UserContext caller = UserContext.current();
-
-        long ipId = lb.getSourceIpAddressId();
-
-        // make sure ip address exists
-        IPAddressVO ipAddr = _ipAddressDao.findById(ipId);
-        if (ipAddr == null || !ipAddr.readyToUse()) {
-            throw new InvalidParameterValueException("Unable to create load balancer rule, invalid IP address id" + ipId);
-        }
-
-        int srcPortStart = lb.getSourcePortStart();
-        int srcPortEnd = lb.getSourcePortEnd();
+    
         int defPortStart = lb.getDefaultPortStart();
         int defPortEnd = lb.getDefaultPortEnd();
 
-        if (!NetUtils.isValidPort(srcPortStart)) {
-            throw new InvalidParameterValueException("publicPort is an invalid value: " + srcPortStart);
-        }
-        if (!NetUtils.isValidPort(srcPortEnd)) {
-            throw new InvalidParameterValueException("Public port range is an invalid value: " + srcPortEnd);
-        }
-        if (srcPortStart > srcPortEnd) {
-            throw new InvalidParameterValueException("Public port range is an invalid value: " + srcPortStart + "-" + srcPortEnd);
-        }
-        if (!NetUtils.isValidPort(defPortStart)) {
-            throw new InvalidParameterValueException("privatePort is an invalid value: " + defPortStart);
-        }
         if (!NetUtils.isValidPort(defPortEnd)) {
             throw new InvalidParameterValueException("privatePort is an invalid value: " + defPortEnd);
         }
@@ -383,30 +373,63 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
         if ((lb.getAlgorithm() == null) || !NetUtils.isValidAlgorithm(lb.getAlgorithm())) {
             throw new InvalidParameterValueException("Invalid algorithm: " + lb.getAlgorithm());
         }
+        
+        LoadBalancer result = _elbMgr.handleCreateLoadBalancerRule(lb, caller.getCaller());
+        if (result == null){
+            result =  createLoadBalancer(lb, openFirewall);
+        } 
+        return result;
+    }
+    
+    @DB
+    public LoadBalancer createLoadBalancer(CreateLoadBalancerRuleCmd lb, boolean openFirewall) throws NetworkRuleConflictException {
+        long ipId = lb.getSourceIpAddressId();
+        UserContext caller = UserContext.current();
+        int srcPortStart = lb.getSourcePortStart();
+        int defPortStart = lb.getDefaultPortStart();
+        int srcPortEnd = lb.getSourcePortEnd();
+        
+        IPAddressVO ipAddr = _ipAddressDao.findById(lb.getSourceIpAddressId());
+        Long networkId = ipAddr.getSourceNetworkId();
+        // make sure ip address exists
+        if (ipAddr == null || !ipAddr.readyToUse()) {
+            throw new InvalidParameterValueException("Unable to create load balancer rule, invalid IP address id" + ipId);
+        }
+        
+        _firewallMgr.validateFirewallRule(caller.getCaller(), ipAddr, srcPortStart, srcPortEnd, lb.getProtocol(), Purpose.LoadBalancing);
 
-        Long networkId = ipAddr.getAssociatedWithNetworkId();
+        
+         networkId = ipAddr.getAssociatedWithNetworkId();
         if (networkId == null) {
             throw new InvalidParameterValueException("Unable to create load balancer rule ; ip id=" + ipId + " is not associated with any network");
 
         }
+        NetworkVO network = _networkDao.findById(networkId);
 
-        _accountMgr.checkAccess(caller.getCaller(), ipAddr);
+        _accountMgr.checkAccess(caller.getCaller(), null, ipAddr);
 
         // verify that lb service is supported by the network
-        Network network = _networkMgr.getNetwork(networkId);
         if (!_networkMgr.isServiceSupported(network.getNetworkOfferingId(), Service.Lb)) {
-            throw new InvalidParameterValueException("LB service is not supported in network id=" + networkId);
+            throw new InvalidParameterValueException("LB service is not supported in network id= " + networkId);
+
         }
 
         Transaction txn = Transaction.currentTxn();
         txn.start();
-
+        
         LoadBalancerVO newRule = new LoadBalancerVO(lb.getXid(), lb.getName(), lb.getDescription(), lb.getSourceIpAddressId(), lb.getSourcePortEnd(), lb.getDefaultPortStart(), 
-                lb.getSourceCidrList(), lb.getAlgorithm(), networkId, ipAddr.getAccountId(), ipAddr.getDomainId());
+                lb.getAlgorithm(), network.getId(), ipAddr.getAccountId(), ipAddr.getDomainId());
+
         newRule = _lbDao.persist(newRule);
+        
+        if (openFirewall) {
+            _firewallMgr.createRuleForAllCidrs(ipId, caller.getCaller(), lb.getSourcePortStart(), lb.getSourcePortEnd(), lb.getProtocol(), null, null, newRule.getId());
+        }
+
+        boolean success = true;
 
         try {
-            _rulesMgr.detectRulesConflict(newRule, ipAddr);
+            _firewallMgr.detectRulesConflict(newRule, ipAddr);
             if (!_rulesDao.setStateToAdd(newRule)) {
                 throw new CloudRuntimeException("Unable to update the state to add for " + newRule);
             }
@@ -415,14 +438,30 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
             UsageEventVO usageEvent = new UsageEventVO(EventTypes.EVENT_LOAD_BALANCER_CREATE, ipAddr.getAllocatedToAccountId(), ipAddr.getDataCenterId(), newRule.getId(), null);
             _usageEventDao.persist(usageEvent);
             txn.commit();
+            
+            // send firewallRule to the backend
+            if (openFirewall) {
+               success = success &&  _firewallMgr.applyFirewallRules(lb.getSourceIpAddressId(), caller.getCaller());
+            }
+            
             return newRule;
         } catch (Exception e) {
-            _lbDao.remove(newRule.getId());
+            success = false;
             if (e instanceof NetworkRuleConflictException) {
                 throw (NetworkRuleConflictException) e;
             }
             throw new CloudRuntimeException("Unable to add rule for ip address id=" + newRule.getSourceIpAddressId(), e);
+        } finally {
+            if (!success) { 
+                txn.start();
+                _firewallDao.remove(_firewallDao.findByRelatedId(newRule.getId()).getId());
+                _lbDao.remove(newRule.getId());
+                txn.commit();
+                _lbDao.remove(newRule.getId());
+            }
         }
+        
+       
     }
 
     @Override
@@ -450,10 +489,8 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
         for (LoadBalancerVO lb : lbs) {
             List<LbDestination> dstList = getExistingDestinations(lb.getId());
 
-            if (dstList != null && !dstList.isEmpty()) {
-                LoadBalancingRule loadBalancing = new LoadBalancingRule(lb, dstList);
-                rules.add(loadBalancing);
-            }
+            LoadBalancingRule loadBalancing = new LoadBalancingRule(lb, dstList);
+            rules.add(loadBalancing);
         }
 
         if (!_networkMgr.applyRules(rules, false)) {
@@ -554,7 +591,7 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
         return true;
     }
 
-    @Override
+    @Override 
     public boolean stop() {
         return true;
     }
@@ -616,7 +653,7 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
             return null;
         }
 
-        _accountMgr.checkAccess(caller, loadBalancer);
+        _accountMgr.checkAccess(caller, null, loadBalancer);
 
         List<UserVmVO> loadBalancerInstances = new ArrayList<UserVmVO>();
         List<LoadBalancerVMMapVO> vmLoadBalancerMappings = null;
